fix arguments reference
When the call to `compute` only has two arguments, some of the
parameters are reassigned. Specifically, the `dependencies` variable is
assigned to the tracking array. This has the unintended effect of
changing `args[1]` to an array (since `arguments[1] === dependencies`)
which will cause a break if there is more than one instance of a view
with that computed property. This patch makes a copy of the arguments
before assigning it to `args`.